Splits out compiled_template
